Product

+++ src/modules/products/application/services/delete-product.service.ts
@@ -1,28 +1,26 @@
 import { Injectable, Inject } from '@nestjs/common';
 import { IProductRepository } from 'src/shared/domain/repository/product-repository.interface';
-import { IProductFactory } from 'src/shared/domain/factories/product-factory.interface';
-import {
-  IProductFactoryToken,
-  IProductRepositoryToken,
-} from 'src/shared/domain/constants';
-import { IDeleteProductService } from 'src/shared/application/services/product-services.interface';
-import { InvalidDatasError } from 'src/modules/users/domain/erros/invalid-data.error';
+import { IProductRepositoryToken } from 'src/shared/domain/constants';
+import { InvalidDatasError } from 'src/shared/domain/erros/invalid-data.error';
 
 @Injectable()
-export class DeleteProductService implements IDeleteProductService {
+export class DeleteProductService {
   constructor(
     @Inject(IProductRepositoryToken)
     private readonly productRepository: IProductRepository,
-
-    @Inject(IProductFactoryToken)
-    private readonly productFactory: IProductFactory,
   ) {}
 
   async deleteProduct(id: string): Promise<void> {
     try {
+      const existingProduct = await this.productRepository.findById(id);
+      if (!existingProduct) {
+        throw new InvalidDatasError();
+      }
       await this.productRepository.delete(id);
-      // eslint-disable-next-line @typescript-eslint/no-unused-vars
     } catch (error) {
+      if (error instanceof InvalidDatasError) {
+        throw new InvalidDatasError();
+      }
       throw new InvalidDatasError();
     }
   }
